:tada: Refactor Pico CMS for Nextcloud
This is a complete code refactoring of how the Nextcloud app handles and passes requests to Pico. It solves all currently known issues with the Nextcloud app and introduces a straight-forward request pipeline.

This was written mostly from scratch, thus there's unfortunately no chance to split this up into smaller commits...

Signed-off-by: Daniel Rudolf <github.com@daniel-rudolf.de>