Add 'Local (unencrypted)' storage backend
The 'Local (unencrypted)' storage backend is registered when server-side encryption is enabled only. It allows admins to still use Pico CMS for Nextcloud even though encryption is enabled, since this storage backend claims that it doesn't support encryption. Use with caution, as it effectively creates an "unencrypted island" in your Nextcloud you should be aware of.

Signed-off-by: Daniel Rudolf <github.com@daniel-rudolf.de>